cdk.json: remove excluding of *.js files
This is for `cdk watch` which will monitor changes to the application code
and redeploy automatically.
Since the lambda code is node-js, remove it from the exclude list.
Now running `cdk watch` in the background will automatically update the lambda code
for us on every save to the lambda function.

I wonder if AWS charges for each lambda update code invokation...

Signed-off-by: Tomer Arbel Eliyahu <tomer.b.eliyahu@intel.com>